+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## ☕️ 2.2.0 (xxxxxxxxxx)
+
+### Bug Fix
+* Fixed a bug with ESlint not being able to tell the difference between node (gulp) and Drupal JS files. This would throw the wrong 'use strict' error.
+
+### New Feature
+* Added a new component subgenerator that can be run as a normal build script `npm run generate`.
+* Updated the start-kit components so they can be easily added to an existing Drupal theme.
+* Added a default `field.html.twig` template that's got less markup in it so that we start new themes in a cleaner state.
+
+### Documentation
+* TODO: Updated the `README.md` with examples of how to use the new component generator and rerun the starter-kit.
+* TODO: Add info about prettier.
+
+## ⚔️ 2.1.0 (2019-11-04)
+
+### Bug Fix
+* Removed patches used by Pattern Lab since the latest release merges in those fixes.
+* Updated readme ESlintrc file link to point to new `.json` file.
+* Updated theme build tool dependencies.
+
+### New Feature
+* Added in support for using the `|clean_id` filter in Pattern Lab the same way you can in Drupal Twig.
+* Migrated eslintrc file from yaml to json.
+
 ## 🏍 2.0.1 (2019-09-27)
 
 ### Bug Fix